Update WS1-API-AdminToolbox_WS1_REST_ConnectionConfig.psm1
Overhaul of the selection options

open-ws1RestConnection - this function can now accept parameters to open the connection. This can allow the function to be used in a non-interactive script.

Additional cleanup has been started for the remainder of the functions to support better error handling and usage of the authType parameters